projects/meme_generator/requirements.txt
autopep8==2.0.4
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
flake8==6.1.0
Flask==2.3.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
mccabe==0.7.0
numpy==1.26.0
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
pip==23.2.1
pluggy==1.3.0
pycodestyle==2.11.0
pyflakes==3.1.0
pytest==7.4.2
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3.post1
requests==2.31.0
setuptools==68.0.0
six==1.16.0
tomli==2.0.1
tzdata==2023.3
urllib3==2.0.5
Werkzeug==2.3.7
wheel==0.41.2
